+++ inbm/integration-reloaded/test/reboot/REBOOT_GOOD_TEST_POSTBOOT.sh
@@ -27,7 +27,7 @@ echo "Waiting for dispatcher to come up postboot..." | systemd-cat
 sleep 3
 echo "Done waiting for dispatcher." | systemd-cat
 
-echo Also looking for /etc/dispatcher_state to be gone
-! [ -f /etc/dispatcher_state ]
+echo Also looking for /var/intel-manageability/dispatcher_state to be gone
+! [ -f /var/intel-manageability/dispatcher_state ]
 
 cleanup_after_test